mcp-modal/requirements.txt
modal
openai
python-dotenv
rich
mcp[cli]
